[CCCL C] Structs use size t for size and alignment members (#3978)
* Modify casting operators of pointer_t and operation_t to initialize all values.

This resolves valgrind error messages about conditional jump based on uninitialized
values shown during run of cccl.c.parallel.test.for.cpp

* Defined all 4 constructors, 1 destructor and 2 assignments for nvrtc_get_name struct

Since the struct member contains a reference, default compiler-generated methods
may lead to undefined behavior. This caused a crash when executing Release binary
of cccl.c.parallel.test.for.cpp test file.

This is output with debugging print statements which since had been removed.

```
executing GetName
executed nvrtcGetLoweredName
==977762== Invalid free() / delete / delete[] / realloc()
==977762==    at 0x484BB6F: operator delete(void*, unsigned long) (in /usr/libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
==977762==    by 0x4877B51: cccl_device_for_build (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/lib/libcccl.c.parallel.so)
==977762==    by 0x12A41E: for_each(cccl_iterator_t, unsigned long, cccl_op_t) (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==    by 0x12DE09: void CATCH2_INTERNAL_TEMPLATE_TEST_0<int>() (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==    by 0x16859E: Catch::RunContext::runCurrentTest() (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==    by 0x168B44: Catch::RunContext::runTest(Catch::TestCaseHandle const&) (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==    by 0x13AB31: Catch::Session::runInternal() (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==    by 0x13AE8D: Catch::Session::run() (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==    by 0x1293F2: main (in /home/coder/cccl/build/cuda12.8-gcc13/cccl-c-parallel/bin/cccl.c.parallel.test.for.cpp)
==977762==  Address 0x1ffeffd5f0 is on thread 1's stack
==977762==  in frame #1, created by cccl_device_for_build (???:)
==977762==
done with GetName
```

After this change the crash disappeared.

Add deleted default constructor for nvrtc_get_name struct

* Change ltsz type from int to size_t, removed static cast

* Implements gh-3960, change size and alignment to be size_t

Size and alignment struct members must have same types as
return types of sizeof(T) and of alignof(T) respectively,
which are both `size_t`.

* Implement gh-3865

Replace unsigned long long types used to describe problem size with uint64_t
to avoid LP64/LLP64 ambiguity.

* Changes _cccl.py to align with type changes in header files

* Changes suggested in review to mem-bound scaling computation in get_policy

This is applies to device_reduce and segmented_device_reduce

* No need to move std::string_view in move constructor of nvrtc_get_name

+++ c/parallel/src/scan.cu
@@ -454,7 +454,7 @@ CUresult cccl_device_scan(
   size_t* temp_storage_bytes,
   cccl_iterator_t d_in,
   cccl_iterator_t d_out,
-  unsigned long long num_items,
+  uint64_t num_items,
   cccl_op_t op,
   cccl_value_t init,
   CUstream stream)
